requirements.txt
langid==1.1.6
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
Pillow==10.0.0
pyspark==3.2.1
seaborn==0.12.2
wordcloud==1.9.2
